requirements.txt
gspread
pandas
openai
faiss-cpu
python-dotenv
fastapi
uvicorn
pydantic 